Fifth commit fixed

+++ 0x14-bit_manipulation/4-clear_bit.c
@@ -4,7 +4,7 @@
  * @n: Pointer to n
  * @index: Index
  */
-int clear_bit(unsigned long int *n, unsigned long int index)
+int clear_bit(unsigned long int *n, unsigned int index)
 {
 	if (index > sizeof(n) * 8)
 		return (-1);
